+++ src/modules/spaces/space.service.ts
@@ -3,12 +3,16 @@ import { UserService } from '../user/user.service'
 import { CreateSpaceDto, UpdateSpaceDto } from './dtos'
 import { SpaceRole } from '../../common/enums'
 import { DbService } from '../../common/db/db.service'
+import { NotificationGateway } from '../notification/notification.gateway'
+import { NotificationType } from '../notification/types'
+import { NotificationService } from '../notification/notification.service'
 
 @Injectable()
 export class SpaceService {
   constructor(
     private readonly db: DbService,
     private readonly userService: UserService,
+    private readonly notificationService: NotificationService,
   ) {}
 
   async createOne(ownerId: string, payload: CreateSpaceDto) {
@@ -23,6 +27,8 @@ export class SpaceService {
       role: SpaceRole.OWNER,
     })
 
+    await this.notificationService.notifyCreateSpace(ownerId, space)
+
     return space
   }
 
@@ -30,17 +36,33 @@ export class SpaceService {
     return this.db.space.findById(spaceId)
   }
 
-  async updateOne(spaceId: string, payload: UpdateSpaceDto) {
-    return this.db.space.findOneAndUpdate(
+  async updateOne(spaceId: string, payload: UpdateSpaceDto, ownerId: string) {
+    const updated = await this.db.space.findOneAndUpdate(
       { _id: spaceId.toString() },
-      {
-        ...payload,
-      },
+      { $set: payload },
       { new: true },
     )
+
+    if (!updated) {
+      throw new NotFoundException(`Space with id ${spaceId} not found`)
+    }
+
+    await this.notificationService.notifyUpdatedSpace(spaceId, ownerId)
+
+    return updated
   }
 
-  async deleteOne(spaceId: string) {
+  async deleteOne(spaceId: string, actor: any) {
+    const space = await this.db.space.findById(spaceId)
+    if (!space)
+      throw new NotFoundException(`Space with id ${spaceId} not found`)
+
+    await this.notificationService.notifyDeletedSpace(
+      spaceId,
+      actor,
+      space.name,
+    )
+
     await this.db.spaceMember.deleteMany({ space: spaceId })
 
     return this.db.space.deleteOne({ _id: spaceId })
